Refractor the threads to separate classes
